refactored out new classes
