+++ Source/EasyNetQ/EasyNetQ.csproj
@@ -130,6 +130,7 @@
     <Compile Include="Interception\IProduceConsumeInterceptor.cs" />
     <Compile Include="Interception\RawMessage.cs" />
     <Compile Include="Interception\TripleDESInterceptor.cs" />
+    <Compile Include="IPersistentDispatcher.cs" />
     <Compile Include="ISaga.cs" />
     <Compile Include="IScheduler.cs" />
     <Compile Include="ISerializer.cs" />
@@ -151,6 +152,7 @@
     <Compile Include="MessageVersioning\MessageVersionStack.cs" />
     <Compile Include="NonGeneric\NonGenericExtensions.cs" />
     <Compile Include="PersistentConnection.cs" />
+    <Compile Include="PersistentDispatcher.cs" />
     <Compile Include="Producer\ClientCommandDispatcher.cs" />
     <Compile Include="Producer\ClientCommandDispatcherSingleton.cs" />
     <Compile Include="Producer\IClientCommandDispatcher.cs" />
