+++ Source/EasyNetQ/IPersistentDispatcher.cs
@@ -1,0 +1,10 @@
+﻿using EasyNetQ.Producer;
+
+namespace EasyNetQ
+{
+    public interface IPersistentDispatcher
+    {
+        IPersistentConnection Connection { get; }
+        IClientCommandDispatcher CommandDispatcher { get; }
+    }
+}
